MQTTClient

Documentation for MQTTClient.

MQTTClient.Client

MQTTClient.MQTTException

MQTTClient.Message

MQTTClient.Packet

MQTTClient.User

MQTTClient.MQTTConnection

MQTTClient.connect_async

MQTTClient.disconnect

MQTTClient.handle_connack

MQTTClient.handle_publish

MQTTClient.keep_alive_loop

MQTTClient.publish

MQTTClient.publish_async

MQTTClient.publish_async

MQTTClient.read_loop

MQTTClient.subscribe

MQTTClient.subscribe_async

MQTTClient.unsubscribe

MQTTClient.unsubscribe_async

MQTTClient.write_loop

Sockets.connect

MQTTClient.@dispatch

MQTTClient.@mqtt_channel

MQTTClient.Client — Type

Client([ping_timeout::UInt64])

A mutable struct that represents an MQTT client.

Arguments

ping_timeout::UInt64: The number of seconds to wait for a ping response before disconnecting. Default is 60.

Fields

on_msg::Dict{String,Function}: A dictionary of functions that will be called when a message is received.

keep_alive::UInt16: The number of seconds between pings.

last_id::UInt16: The last message ID used.

in_flight::Dict{UInt16, Future}: A dictionary of messages that are waiting for a response.

write_packets::AbstractChannel: A channel for writing packets to the socket.

socket: The socket used for communication.

socket_lock: A lock for the socket.

ping_timeout::UInt64: The number of seconds to wait for a ping response before disconnecting.

ping_outstanding::Atomic{UInt8}: A flag indicating whether a ping has been sent but not yet received.

last_sent::Atomic{Float64}: The time the last packet was sent.

last_received::Atomic{Float64}: The time the last packet was received.

Examples

julia> client = Client()

MQTTClient.MQTTException — Type

struct MQTTException <: Exception
    msg::AbstractString
end

A custom exception type for MQTT errors.

# Examples
```julia-repl
julia> throw(MQTTException("Connection refused: Not authorized"))
MQTTException("Connection refused: Not authorized")
```

MQTTClient.Message — Type

Message

A composite type representing a message.

Fields

dup::Bool: a boolean indicating whether the message is a duplicate.

qos::UInt8: an 8-bit unsigned integer representing the quality of service.

retain::Bool: a boolean indicating whether the message should be retained.

topic::String: a string representing the topic.

payload::Array{UInt8}: an array of 8-bit unsigned integers representing the payload.

Constructors

Message(qos::QOS, topic::String, payload...): constructs a new message with default values for dup and retain.

Message(dup::Bool, qos::QOS, retain::Bool, topic::String, payload...): constructs a new message with all fields specified.

Message(dup::Bool, qos::UInt8, retain::Bool, topic::String, payload...): constructs a new message with all fields specified.

MQTTClient.Packet — Type

Packet

A composite type representing a packet.

Fields

cmd::UInt8: an 8-bit unsigned integer representing the command.

data::Any: any value representing the data.

MQTTClient.User — Type

User(name::String, password::String)

A struct that represents a user with a name and password.

Examples

julia> user = User("John", "password")
User("John", "password")

MQTTClient.MQTTConnection — Method

MQTTConnection(;ping_timeout=UInt64(60))

Create a new Client object with the specified ping_timeout (optional).

Keyword Arguments

ping_timeout::UInt64=60: The number of seconds to wait for a ping response before disconnecting. Default is 60.

Examples

client = MQTTConnection()
client = MQTTConnection(ping_timeout=30)

MQTTClient.connect_async — Function

connect_async(client::Client, host::AbstractString, port::Integer=1883;
   keep_alive::UInt16=0x0000,
   client_id::String=randstring(8),
   user::User=User("", ""),
   will::Message=Message(false, 0x00, false, "", Array{UInt8}()),
   clean_session::Bool=true)

Connects the Client instance to the specified broker. Returns a Future object that contains a session_present bit from the broker on success and an exception on failure.

Arguments

keep_alive::Int64=0: Time in seconds to wait before sending a ping to the broker if no other packets are being sent or received.

client_id::String=randstring(8): The id of the client.

user::User=User("", ""): The MQTT authentication.

will::Message=Message(false, 0x00, false, "", Array{UInt8}()): The MQTT will to send to all other clients when this client disconnects.

clean_session::Bool=true: Flag to resume a session with the broker if present.

MQTTClient.disconnect — Method

disconnect(client::Client)

Disconnects the client from the broker and stops the tasks.

MQTTClient.handle_connack — Method

handle_connack(client::Client, s::IO, cmd::UInt8, flags::UInt8)

Handle a CONNACK packet.

Arguments

client::Client: The client that received the packet.

s::IO: The socket used for communication.

cmd::UInt8: The command byte.

flags::UInt8: The flags byte.

Examples

julia> handle_connack(client, s, 0x20, 0x00)

MQTTClient.handle_publish — Method

handle_publish(client::Client, s::IO, cmd::UInt8, flags::UInt8)

This function handles the publish command in MQTT protocol.

Arguments

client: A client object.

s: An input stream.

cmd: An unsigned 8-bit integer.

flags: An unsigned 8-bit integer.

Returns

Future.

MQTTClient.keep_alive_loop — Method

keep_alive_loop(client::Client)

This function runs a loop that sends a PINGREQ message to the MQTT broker to keep the connection alive. The loop checks the connection at regular intervals determined by the client.keep_alive value. If no message has been sent or received within the keep-alive interval, a PINGREQ message is sent. If no PINGRESP message is received within the client.ping_timeout interval, the client is disconnected.

MQTTClient.publish — Method

publish(client::Client, topic::String, payload...; dup::Bool=false, qos::QOS=QOS_0, retain::Bool=false)

Waits until the publish is completely acknowledged. Publishes a message with the specified parameters. Returns nothign on success and throws an exception on failure.

MQTTClient.publish_async — Method

publish_async(client::Client, message::Message)

Publishes the message. Returns a Future object that contains nothing on success and an exception on failure.

MQTTClient.publish_async — Method

publish_async(client::Client, topic::String, payload...;
   dup::Bool=false,
   qos::QOS=QOS_0,
   retain::Bool=false)

Pulishes a message with the specified parameters. Returns a Future object that contains nothing on success and an exception on failure.

MQTTClient.read_loop — Method

read_loop(client)

Reads data from a client socket and processes it.

Arguments

client: A client object.

Example

read_loop(client)

MQTTClient.subscribe — Method

subscribe(client::Client, topic::String, on_msg::Function; qos::UInt8=QOS_0)

Subscribe to a topic.

Arguments

client::Client: The MQTT client.

topic::String: The topic to subscribe to.

on_msg::Function: The function to call when a message is received on the topic.

qos::UInt8: The quality of service level to use for the subscription. Default is 0.

Examples

subscribe(client, "my/topic", on_msg)

MQTTClient.subscribe_async — Method

subscribe_async(client::Client, topic::String, on_msg::Function; qos::UInt8=QOS_0)

Subscribe to a topic asynchronously.

Arguments

client::Client: The MQTT client.

topic::String: The topic to subscribe to.

on_msg::Function: The function to call when a message is received on the topic.

qos::UInt8: The quality of service level to use for the subscription. Default is 0.

Returns

Future: A future that can be used to wait for the subscription to complete.

Examples

future = subscribe_async(client, "my/topic", on_msg, qos=QOS_2)

MQTTClient.unsubscribe — Method

unsubscribe(client::Client, topics::String...)

Unsubscribes the Client instance from the supplied topic names. Waits until the unsubscribe is fully acknowledged. Returns nothing on success and an exception on failure.

MQTTClient.unsubscribe_async — Method

unsubscribe_async(client::Client, topics::String...)

Unsubscribes the Client instance from the supplied topic names. Deletes the callback from the client Returns a Future object that contains nothing on success and an exception on failure.

MQTTClient.write_loop — Method

write_loop(client)

This function writes data to the socket.

Arguments

client: A client object.

Returns

Nothing.

Sockets.connect — Function

connect(client::Client, host::AbstractString, port::Integer=1883;
    keep_alive::UInt16=0x0000,
    client_id::String=randstring(8),
    user::User=User("", ""),
    will::Message=Message(false, 0x00, false, "", Array{UInt8}()),
    clean_session::Bool=true)

Connects the Client instance to the specified broker. Waits until the connect is done. Returns the session_present bit from the broker on success and an exception on failure.

Arguments

keep_alive::Int64=0: Time in seconds to wait before sending a ping to the broker if no other packets are being sent or received.

client_id::String=randstring(8): The id of the client.

user::User=User("", ""): The MQTT authentication.

will::Message=Message(false, 0x00, false, "", Array{UInt8}()): The MQTT will to send to all other clients when this client disconnects.

clean_session::Bool=true: Flag to resume a session with the broker if present.

MQTTClient.@dispatch — Macro

@dispatch(ex)

A macro that dispatches the execution of an expression ex asynchronously.

If the number of threads is equal to 1, it uses the @async macro to execute the expression asynchronously.

If the number of threads is greater than 1, it uses the Dagger.@spawn macro to execute the expression asynchronously.

If the number of threads is not valid, it throws an exception.

Examples

julia> @dispatch println("Hello, World!")
Task (done) @0x00007f8c3e8a1010

julia> @dispatch begin
           sleep(1)
           println("Hello, World!")
       end
Task (done) @0x00007f8c3e8a1010

MQTTClient.@mqtt_channel — Macro

mqtt_channel(len::Number=128)

A macro that declares a data channel based on the number of threads available. If more than one thread is available, it returns a RemoteChannel with a Channel{Packet} of length len. Otherwise, it returns a Channel{Packet} of length len.

Arguments

len::Number=128: The length of the channel. Defaults to 128.

Examples

@mqtt_channel  # Returns a Channel{Packet} of length 128
@mqtt_channel 64  # Returns a Channel{Packet} of length 64