MQTTClient.MakeConnection — Function

MakeConnection(host::Union{IPAddr, String}, port::Int64;
    ping_timeout=UInt64(60),
    keep_alive::Int64=32,
    client_id::String=randstring(8),
    user::User=User("", ""),
    will::Message=Message(false, 0x00, false, "", UInt8[]),
    clean_session::Bool=true) -> Tuple{Client, MQTTConnection}

Establishes a connection to the given host and port. The host can be either an IP address or a hostname.

Arguments

host: The IP address or hostname to connect to.

port: The port number to connect to.

ping_timeout: The maximum time in seconds to wait for a ping response before considering the connection lost.

keep_alive: The maximum time in seconds that the client should wait before sending a ping request to keep the connection alive.

client_id: The client identifier string.

user: The user credentials for authentication.

will: The last will and testament message to be sent by the server on behalf of the client.

clean_session: Whether to start a clean session or resume a previous one.

Returns

A tuple containing a Client object and an MQTTConnection object.

MakeConnection(path::String;
    ping_timeout=UInt64(60),
    keep_alive::Int64=32,
    client_id::String=randstring(8),
    user::User=User("", ""),
    will::Message=Message(false, 0x00, false, "", UInt8[]),
    clean_session::Bool=true)  -> Tuple{Client, MQTTConnection}

Establishes a connection to a Unix domain socket at the given path.

Arguments

path: The path of the Unix domain socket to connect to.

ping_timeout: The maximum time in seconds to wait for a ping response before considering the connection lost.

keep_alive: The maximum time in seconds that the client should wait before sending a ping request to keep the connection alive.

client_id: The client identifier string.

user: The user credentials for authentication.

will: The last will and testament message to be sent by the server on behalf of the client.

clean_session: Whether to start a clean session or resume a previous one.

Returns

A tuple containing a Client object and an MQTTConnection object.

MakeConnection(io::T,
    ping_timeout=UInt64(60),
    keep_alive::Int64=32,
    client_id::String=randstring(8),
    user::User=User("", ""),
    will::Message=Message(false, 0x00, false, "", UInt8[]),
    clean_session::Bool=true)  -> Tuple{Client, MQTTConnection}

Establishes an MQTT connection using the given IO connection.

Arguments

io: An IO connection object that implements the AbstractIOConnection interface.

ping_timeout: The maximum time in seconds to wait for a ping response before considering the connection lost.

keep_alive: The maximum time in seconds that the client should wait before sending a ping request to keep the connection alive.

client_id: The client identifier string.

user: The user credentials for authentication.

will: The last will and testament message to be sent by the server on behalf of the client.

clean_session: Whether to start a clean session or resume a previous one.

Returns

A tuple containing a Client object and an MQTTConnection object.

MQTTClient.connect_async — Function

connect_async(client::Client, connection::MQTTConnection)

Establishes an asynchronous connection to an MQTT broker using the specified client and connection objects.

The function sets up the write and read loops, as well as the keep-alive loop if the keep-alive time is greater than 0. It also sets the protocol name and level, and the connect flags.

If the user name and password are provided in the connection object, they are included in the connect flags. If a will topic is provided, it is also included in the connect flags.

The function returns a Future object that can be used to track the progress of the connection.

Sockets.connect — Function

connect(protocol::UDS) -> PipeEndpoint

Establishes a connection to a Unix domain socket at the given path specified in the UDS struct.

connect(protocol::TCP) -> TCPSocket

Establishes a TCP connection to the given IP address and port specified in the TCP struct.

connect(client::Client, connection::MQTTConnection)

Establishes a synchronous connection to an MQTT broker using the specified client and connection objects.

This function is a wrapper around the connect_async function, which establishes an asynchronous connection. The connect function blocks until the connection is established by calling the resolve function on the Future object returned by connect_async.

MQTTClient.disconnect — Function

disconnect(client::Client)

Disconnects the client from the broker and stops the tasks.

MQTTClient.subscribe_async — Function

subscribe_async(client::Client, topic::String, on_msg::Function; qos::UInt8=QOS_0)

Subscribe to a topic asynchronously.

Arguments

client::Client: The MQTT client.

topic::String: The topic to subscribe to.

on_msg::Function: The function to call when a message is received on the topic.

qos::UInt8: The quality of service level to use for the subscription. Default is 0.

Returns

Future: A future that can be used to wait for the subscription to complete.

Examples

future = subscribe_async(client, "my/topic", on_msg, qos=QOS_2)

MQTTClient.subscribe — Function

subscribe(client::Client, topic::String, on_msg::Function; qos::UInt8=QOS_0)

Subscribe to a topic.

Arguments

client::Client: The MQTT client.

topic::String: The topic to subscribe to.

on_msg::Function: The function to call when a message is received on the topic.

qos::UInt8: The quality of service level to use for the subscription. Default is 0.

Examples

subscribe(client, "my/topic", on_msg)

MQTTClient.unsubscribe_async — Function

unsubscribe_async(client::Client, topics::String...)

Unsubscribes the Client instance from the supplied topic names. Deletes the callback from the client Returns a Future object that contains nothing on success and an exception on failure.

MQTTClient.unsubscribe — Function

unsubscribe(client::Client, topics::String...)

Unsubscribes the Client instance from the supplied topic names. Waits until the unsubscribe is fully acknowledged. Returns nothing on success and an exception on failure.

MQTTClient.publish_async — Function

publish_async(client::Client, message::Message)

Publishes the message. Returns a Future object that contains nothing on success and an exception on failure.

publish_async(client::Client, topic::String, payload...;
   dup::Bool=false,
   qos::QOS=QOS_0,
   retain::Bool=false)

Pulishes a message with the specified parameters. Returns a Future object that contains nothing on success and an exception on failure.

MQTTClient.publish — Function

publish(client::Client, topic::String, payload...; dup::Bool=false, qos::QOS=QOS_0, retain::Bool=false)

Waits until the publish is completely acknowledged. Publishes a message with the specified parameters. Returns nothign on success and throws an exception on failure.